Set special diff tool for .ipynb files
`.ipynb` files are not easy to diff without special tools. To work
around this problem, I decided to use `nbdime` which has a special
command to configure git to use its own tool for diffing. The command
that needs to be executed is:

```
nbdime config-git --enable # add --global for global configuration
```

Alternatively, the following can be added to the local (`.git/config`)
or global (`~/.gitconfig`) configuration:

```
[diff "jupyternotebook"]
	command = git-nbdiffdriver diff
[merge "jupyternotebook"]
	driver = git-nbmergedriver merge %O %A %B %L %P
	name = jupyter notebook merge driver
[difftool "nbdime"]
	cmd = git-nbdifftool diff \"$LOCAL\" \"$REMOTE\" \"$BASE\"
[difftool]
	prompt = false
[mergetool "nbdime"]
	cmd = git-nbmergetool merge \"$BASE\" \"$LOCAL\" \"$REMOTE\" \"$MERGED\"
[mergetool]
	prompt = false
```